% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-temporal-measurement-error.R
\name{apply_anticipatory_measurement}
\alias{apply_anticipatory_measurement}
\title{Helper: Apply anticipatory measurement}
\usage{
apply_anticipatory_measurement(data, tme)
}
\description{
Helper: Apply anticipatory measurement
}
\keyword{internal}
