% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-temporal-measurement-error.R
\name{apply_delayed_measurement}
\alias{apply_delayed_measurement}
\title{Helper: Apply delayed measurement}
\usage{
apply_delayed_measurement(data, tme)
}
\description{
Helper: Apply delayed measurement
}
\keyword{internal}
