% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-temporal-measurement-error.R
\name{apply_retrospective_error}
\alias{apply_retrospective_error}
\title{Helper: Apply retrospective measurement error}
\usage{
apply_retrospective_error(data, tme, true_data = NULL)
}
\description{
Helper: Apply retrospective measurement error
}
\keyword{internal}
