% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-panel.R
\name{as_wide.margot_panel}
\alias{as_wide.margot_panel}
\title{Convert margot_panel to wide format}
\usage{
as_wide.margot_panel(x, sep = "_", ...)
}
\arguments{
\item{x}{A margot_panel object}

\item{sep}{Character separator for column names (default: "_")}

\item{...}{Additional arguments passed to reshape}
}
\value{
Wide format data frame
}
\description{
Convert margot_panel to wide format
}
