% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-core-utils.R
\name{.margot_core_generics}
\alias{.margot_core_generics}
\title{Get list of generic functions in margot.core}
\usage{
.margot_core_generics()
}
\value{
Character vector of generic function names
}
\description{
Get list of generic functions in margot.core
}
