% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-panel.R
\name{is_margot_panel}
\alias{is_margot_panel}
\title{Check if object is a margot_panel}
\usage{
is_margot_panel(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Check if object is a margot_panel
}
