% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-scm-enhanced.R
\name{is_margot_scm}
\alias{is_margot_scm}
\title{Check if object is a margot_scm}
\usage{
is_margot_scm(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if object is a margot_scm
}
