% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-scm.R
\name{is_scm}
\alias{is_scm}
\title{Check if object is an SCM}
\usage{
is_scm(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if x is an scm object
}
\description{
Check if object is an SCM
}
