% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-core-utils.R
\name{margot_option_defaults}
\alias{margot_option_defaults}
\title{Get default margotsphere options}
\usage{
margot_option_defaults()
}
\value{
List of default option values
}
\description{
Get default margotsphere options
}
