% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-scm-enhanced.R
\name{margot_scm_add_shift}
\alias{margot_scm_add_shift}
\title{Add shift intervention to SCM}
\usage{
margot_scm_add_shift(scm, node, name, fun)
}
\arguments{
\item{scm}{A margot_scm object}

\item{node}{Name of node to intervene on}

\item{name}{Name for this intervention}

\item{fun}{Shift function}
}
\value{
Updated margot_scm object
}
\description{
Add shift intervention to SCM
}
\examples{
scm <- margot_scm_from_dsl("
U = U_a, U_y
k1 = a ~ U_a
k2 = y ~ a + U_y
")
scm <- margot_scm_add_shift(scm, "a", "increase", function(a, ...) a + 1)
}
