% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-scm-enhanced.R
\name{margot_scm_as_latex}
\alias{margot_scm_as_latex}
\title{Convert margot_scm to LaTeX}
\usage{
margot_scm_as_latex(x)
}
\arguments{
\item{x}{A margot_scm object}
}
\value{
LaTeX string
}
\description{
Convert margot_scm to LaTeX
}
