% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-scm-enhanced.R
\name{margot_scm_from_df}
\alias{margot_scm_from_df}
\title{Build SCM from data frame with automatic exogenous mapping}
\usage{
margot_scm_from_df(data, K_min = 3)
}
\arguments{
\item{data}{Data frame with LMTP naming convention}

\item{K_min}{Minimum number of time periods required}
}
\value{
Object of class \code{margot_scm}
}
\description{
Build SCM from data frame with automatic exogenous mapping
}
