% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-scm-enhanced.R
\name{margot_scm_time_indices}
\alias{margot_scm_time_indices}
\title{Get time indices from SCM}
\usage{
margot_scm_time_indices(scm)
}
\arguments{
\item{scm}{A margot_scm object}
}
\value{
Named vector of time indices
}
\description{
Get time indices from SCM
}
