% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-panel.R
\name{panel_attrs}
\alias{panel_attrs}
\title{Extract panel attributes}
\usage{
panel_attrs(x)
}
\arguments{
\item{x}{A margot_panel object}
}
\value{
List with id_col and time_col
}
\description{
Extract panel attributes
}
