% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-panel.R
\name{print.margot_panel}
\alias{print.margot_panel}
\title{Print method for margot_panel objects}
\usage{
\method{print}{margot_panel}(x, ...)
}
\arguments{
\item{x}{A margot_panel object}

\item{...}{Additional arguments passed to print methods}
}
\value{
The object invisibly
}
\description{
Print method for margot_panel objects
}
