% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-scm-enhanced.R
\name{print.margot_scm}
\alias{print.margot_scm}
\title{Print method for margot_scm objects}
\usage{
\method{print}{margot_scm}(x, ..., indent = "  ", width = getOption("width", 80))
}
\arguments{
\item{x}{A margot_scm object}

\item{...}{Additional arguments}

\item{indent}{Indentation string}

\item{width}{Line width}
}
\value{
Invisibly returns the object
}
\description{
Print method for margot_scm objects
}
