% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-temporal-measurement-error.R
\name{print.margot_temporal_measurement_error}
\alias{print.margot_temporal_measurement_error}
\title{Print method for temporal measurement error}
\usage{
\method{print}{margot_temporal_measurement_error}(x, ...)
}
\description{
Print method for temporal measurement error
}
