% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-timeline.R
\name{print.margot_timeline}
\alias{print.margot_timeline}
\title{Print method for margot_timeline objects}
\usage{
\method{print}{margot_timeline}(x, ...)
}
\arguments{
\item{x}{A margot_timeline object}

\item{...}{Additional arguments passed to print}
}
\value{
The object invisibly
}
\description{
Print method for margot_timeline objects
}
