% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-scm.R
\name{print.scm}
\alias{print.scm}
\title{Print method for SCM objects}
\usage{
\method{print}{scm}(x, indent = "  ", width = getOption("width", 80), ...)
}
\arguments{
\item{x}{An scm object}

\item{indent}{Indentation string (default "  ")}

\item{width}{Maximum line width (default from options)}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the SCM object
}
\description{
Print method for SCM objects
}
