% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{print.temporal_measurement_error_shadow}
\alias{print.temporal_measurement_error_shadow}
\title{Print temporal measurement error shadow}
\usage{
\method{print}{temporal_measurement_error_shadow}(x, ...)
}
\arguments{
\item{x}{A temporal_measurement_error_shadow object}

\item{...}{Additional arguments}
}
\value{
The object invisibly
}
\description{
Print temporal measurement error shadow
}
