% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-seed.R
\name{restore_seed_state}
\alias{restore_seed_state}
\title{Restore RNG State}
\usage{
restore_seed_state(state)
}
\arguments{
\item{state}{RNG state from preserve_seed()}
}
\value{
NULL (invisibly)
}
\description{
Restores a previously saved RNG state.
}
