% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-scm.R
\name{scm_from_spec}
\alias{scm_from_spec}
\title{Build an SCM from equation specifications}
\usage{
scm_from_spec(spec)
}
\arguments{
\item{spec}{A single character string with equations separated by semicolons}
}
\value{
Object of class \code{scm}
}
\description{
Creates an SCM object from a semicolon-separated string of equations.
}
\examples{
spec <- "t0_b ~ U1 + U2; t1_a ~ U1; t2_y ~ U2"
scm <- scm_from_spec(spec)
print(scm)
}
