% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-scm.R
\name{summary.scm}
\alias{summary.scm}
\title{Summary method for SCM objects}
\usage{
\method{summary}{scm}(object, ...)
}
\arguments{
\item{object}{An scm object}

\item{...}{Additional arguments (unused)}
}
\value{
Summary information about the SCM
}
\description{
Summary method for SCM objects
}
