% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-scm.R
\name{validate_scm}
\alias{validate_scm}
\title{Validate SCM object}
\usage{
validate_scm(x)
}
\arguments{
\item{x}{An scm object}
}
\value{
TRUE if valid, error otherwise
}
\description{
Validate SCM object
}
\keyword{internal}
