% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-temporal-measurement-error.R
\name{validate_temporal_measurement_error}
\alias{validate_temporal_measurement_error}
\title{Validate temporal measurement error object}
\usage{
validate_temporal_measurement_error(x)
}
\arguments{
\item{x}{A margot_temporal_measurement_error object}
}
\description{
Validate temporal measurement error object
}
