% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-timeline.R
\name{validate_timeline}
\alias{validate_timeline}
\title{Validate a timeline object}
\usage{
validate_timeline(x, ...)
}
\arguments{
\item{x}{A timeline object to validate}

\item{...}{Additional arguments (unused)}
}
\value{
The validated timeline object (invisibly)
}
\description{
Validate a timeline object
}
